# [OptimizeInstructions] Improve cache misses for canonicalize routines