--- a/src/passes/OptimizeInstructions.cpp
+++ b/src/passes/OptimizeInstructions.cpp
@@ -277,7 +277,7 @@ struct OptimizeInstructions
     // If this contains dead code, don't bother trying to optimize it, the type
     // might change (if might not be unreachable if just one arm is, for
     // example). This optimization pass focuses on actually executing code.
-    if (curr->type == Type::unreachable) {
+    if (LLVM_UNLIKELY(curr->type == Type::unreachable)) {
       return;
     }