--- a/src/passes/OptimizeInstructions.cpp
+++ b/src/passes/OptimizeInstructions.cpp
@@ -1848,7 +1848,7 @@ private:
       ZeroRemover(PassOptions& passOptions) : passOptions(passOptions) {}
 
       void visitBinary(Binary* curr) {
-        if (!curr->type.isInteger()) {
+        if (WASM_UNLIKELY(!curr->type.isInteger())) {
           return;
         }
         FeatureSet features = getModule()->features;