Usually after first OptimizeInstructions pass we already have pretty canonicalization IR which is rarely violated by next passes, so we can improve caching by marking conditions as unlikely in canonicalize routines.