@MaxGraey 2021-08-10
### Benchmarks
`wasm-opt` for 1.5 mb input real world file.

#### main branch:

1) [PassRunner]   running pass: optimize-instructions... 0.0632812 seconds.
2) [PassRunner]   running pass: optimize-instructions... 0.0586864 seconds.
3) [PassRunner]   running pass: optimize-instructions... 0.0608243 seconds.

####  this PR branch:  
1) [PassRunner]   running pass: optimize-instructions... 0.0535724 seconds.
2) [PassRunner]   running pass: optimize-instructions... 0.0558105 seconds.
3) [PassRunner]   running pass: optimize-instructions... 0.0529371 seconds.
4) [PassRunner]   running pass: optimize-instructions... 0.0551918 seconds.