@kripken 2021-08-10
Interesting... this may be worth doing. Can you measure on some large things, though?

If we do this, we should probably add `BYN_UNLIKELY` as we don't want to depend on LLVM code here.

Most generally, though, it seems like every single optimization pattern in this file is unlikely - probably it finds one in a hundred or so (just a guess). So I wonder if we shouldn't annotate them all somehow. But then that would probably remove the benefit. Maybe it should be only for unreachability and things like that.