@MaxGraey 2021-08-11
Here another one for 4 runs. Same file (1.5 mb unoptimized wasm file)

main with -O3:

1. [PassRunner] passes took 4.17425 seconds.
2. [PassRunner] passes took 4.218 seconds.
3. [PassRunner] passes took 4.19139 seconds.
4. [PassRunner] passes took 4.15892 seconds.


this PR with -O3:

  1. [PassRunner] passes took 4.12122 seconds.
2. [PassRunner] passes took 4.09263 seconds.
3. [PassRunner] passes took 4.14114 seconds.
4. [PassRunner] passes took 4.09389 seconds.