@kripken 2021-08-16
I made an LTO build and tested this on some large wasm files. I actually don't see a speedup here. Maybe that's due to my CPU or some other difference?