@MaxGraey 2021-08-17
Perhaps it depends on what is wasm module. I assume these modules were generated by LLVM? It's just that in this case OptimizeInstruction will rarely be called I think