@kripken 2021-08-17
One was from LLVM, the other was a very big one from J2CL (Wasm GC from Java). But maybe you're right that both use this pass less. Do you have a link to a large AssemblyScript one perhaps?