@MaxGraey 2021-08-17
Yes, I used this file:
[assemblyscript.untouched.wasm.zip](https://github.com/WebAssembly/binaryen/files/7001920/assemblyscript.untouched.wasm.zip)