@kripken 2021-08-23
Ok, I tested a lot on that file, both doing `--optimize-instructions` by itself as well as `-O1` and `-O3` which runs it internally. I don't see a difference. Maybe certain CPUs benefit and others don't...

Maybe an argument could be made that this improves code clarity, by marking what is likely and what not. But I'm not sure that it does, myself.