@MaxGraey 2021-08-24
The acceleration is really very slight and will depend on the size of the cache. Well I can close this PR. Or is it better to leave only the new macros but delete everything else?