@kripken 2021-08-24
I think let's close it for now. The macros do look useful, but we can return to this PR when we find their first user.